refactor: move array and array stacked to subfolder
