+++ docarray/__init__.py
@@ -1,6 +1,6 @@
 __version__ = '2023.01.18.alpha'
 
-from docarray.array.array import DocumentArray
+from docarray.array.array.array import DocumentArray
 from docarray.base_document.document import BaseDocument
 
 __all__ = [
