+++ docarray/utils/find.py
@@ -3,8 +3,8 @@ from typing import List, NamedTuple, Optional, Type, Union, cast
 from typing_inspect import is_union_type
 
 from docarray.array.abstract_array import AnyDocumentArray
-from docarray.array.array import DocumentArray
-from docarray.array.array_stacked import DocumentArrayStacked
+from docarray.array.array.array import DocumentArray
+from docarray.array.stacked.array_stacked import DocumentArrayStacked
 from docarray.base_document import BaseDocument
 from docarray.typing import AnyTensor
 from docarray.typing.tensor.abstract_tensor import AbstractTensor
